+++ src/widgets/progressbar.ts
@@ -1,7 +1,5 @@
 import AgsWidget, { type BaseProps } from './widget.js';
-import GObject from 'gi://GObject';
 import Gtk from 'gi://Gtk?version=3.0';
-import Service from '../service.js';
 
 export type ProgressBarProps = BaseProps<AgsProgressBar, Gtk.ProgressBar.ConstructorProperties & {
     vertical?: boolean
@@ -10,13 +8,12 @@ export type ProgressBarProps = BaseProps<AgsProgressBar, Gtk.ProgressBar.Constru
 
 export default class AgsProgressBar extends AgsWidget(Gtk.ProgressBar) {
     static {
-        GObject.registerClass({
-            GTypeName: 'AgsProgressBar',
-            Properties: {
-                'vertical': Service.pspec('vertical', 'boolean', 'rw'),
-                'value': Service.pspec('value', 'float', 'rw'),
+        AgsWidget.register(this, {
+            properties: {
+                'vertical': ['boolean', 'rw'],
+                'value': ['float', 'rw'],
             },
-        }, this);
+        });
     }
 
     constructor(props: ProgressBarProps = {}) {